Replaced CoinbaseWebSocket with the state machine one
